Prefer 'atom-workspace atom-text-editor' rule
Based on the Linux keymaps[1] and the Darwin keymaps[2] it looks like we
need to put most atom-text-editor rules inside the 'atom-workspace
atom-text-editor' section.  That's where they are on Linux and our rules
need to be at least as specific as the default rules to take effect.
This makes our rules more specific than Darwin's default, so we're OK
there too.

1: https://github.com/atom/atom/blob/daca18f2e533af66ebe3bd90db2d4e0a541c8bfe/keymaps/linux.cson
2: https://github.com/atom/atom/blob/daca18f2e533af66ebe3bd90db2d4e0a541c8bfe/keymaps/linux.cson